+++ R/ddfilter.r
@@ -23,8 +23,10 @@
 #' or to a subsequent location exceeds a given "maxvlp". If "vmax" and "maxvlp" are unknown, they can be estimated 
 #' using the functions "est.vmax" and "est.maxvlp", respectively.
 #' @return A data frame is returned without locations identified by this filter. 
-#' The following columns are added: "pTime", "pDist", "inAng". 
-#' "pTime" is hours from a previous fix. "pDist" is straight distance in kilometres from a previous fix. 
+#' The following columns are added: "pTime", "sTime", "pDist", "sDist", "pSpeed", "sSpeed", "inAng". 
+#' "pTime" and "sTime" are hours from a previous and to a subsequent fix respectively. 
+#' "pDist" and "sDist" are straight distances in kilometres from a previous and to a subsequent fix respectively. 
+#' "pSpeed" and "sSpeed" are linear speed from a previous and to a subsequent fix respectively. 
 #' "inAng" is the angle between the bearings of lines joining successive location points.    
 #' @author Takahiro Shimada
 #' @references Shimada T, Jones R, Limpus C, Hamann M (2012)