+++ R/distancefilter.r
@@ -1,40 +1,32 @@
-#' @aliases ddfilter.speed
-#' @title Filter locations by speed 
-#' @description A partial component of ddfilter, although works as a stand-alone function. 
-#' This function removes locations by a given threshold speed as described in Shimada et al. (2012).
+#' @aliases distance_filter
+#' @title Filter locations by distance 
+#' @description This function removes locations that are located beyond a given threshold distance.
 #' @param sdata A data frame containing columns with the following headers: "id", "DateTime", "lat", "lon", "qi". 
 #' This filter is independently applied to a subset of data grouped by the unique "id". 
 #' "DateTime" is date & time in class \code{\link[base]{POSIXct}}. "lat" and "lon" are the recorded latitude and longitude in decimal degrees. 
 #' "qi" is the numerical quality index associated with each fix where the greater number represents better quality 
 #' (e.g. number of GPS satellites used for estimation).
-#' @param vmax A numeric value specifying threshold speed both from a previous and to a subsequent fix. Default is 8.9 km/h. 
-#' If this value is unknown, the function "est.vmax" can be used to estimate the value based on the supplied data.
-#' @param method An integer specifying how locations are filtered by speed. 
-#' 1 = a location is removed if the speed EITHER from a previous and to a subsequent location exceeds a given threshold speed. 
-#' 2 = a location is removed if the speed BOTH from a previous and to a subsequent location exceeds a given threshold speed. Default is FALSE.
+#' @param distance1 A numeric value specifying threshold distance from a previous fix. Default is 100 km. 
+#' @param distance2 A numeric value specifying threshold distance to a subsequent fix. Default is 100 km. 
+#' @param method An integer specifying how locations are filtered by distance. 
+#' 1 = a location is removed if the distance EITHER from a previous and to a subsequent location exceeds a given threshold speed. 
+#' 2 = a location is removed if the distance BOTH from a previous and to a subsequent location exceeds a given threshold speed. Default is FALSE.
 #' @import sp
 #' @importFrom raster pointDistance
 #' @export
-#' @details This function removes locations if the speed both/either from a previous and to a subsequent location exceeds a given threshold speed. 
-#' If "vmax" is unknown, it can be estimated using the function "est.vmax".
+#' @details This function removes locations if the distance both/either from a previous and to a subsequent location exceeds a given threshold distance. 
 #' @return A data frame is returned without locations identified by this filter. 
-#' The following columns are added: "pTime", "sTime", "pDist", "sDist", "pSpeed", "sSpeed". 
-#' "pTime" and "sTime" are hours from a previous and to a subsequent fix respectively. 
-#' "pDist" and "sDist" are straight distances in kilometres from a previous and to a subsequent fix respectively. 
-#' "pSpeed" and "sSpeed" are linear speed from a previous and to a subsequent fix respectively. 
+#' The following columns are added: "pTime", "pDist". 
+#' "pTime" is hours from a previous fix. "pDist" is straight distance in kilometres from a previous fix. 
 #' @author Takahiro Shimada
-#' @references Shimada T, Jones R, Limpus C, Hamann M (2012) 
-#' Improving data retention and home range estimates by data-driven screening.
-#' Marine Ecology Progress Series 457:171-180 doi:10.3354/meps09747
-#' @seealso \code{\link{ddfilter}}, \code{\link{ddfilter.loop}}, \code{\link{est.vmax}}
 
 
 
-ddfilter.speed<-function (sdata, vmax=8.9, method=2){
+distance_filter<-function (sdata, distance1=100, distance2=100, method=2){
   
   OriginalSS<-nrow(sdata)
     
-    max.speed<-function(sdata, vmax, method){
+    max.distance<-function(sdata, distance1, distance2, method){
       #### Exclude data with less than 4 locations
       ndata<-table(sdata$id)
       id.exclude<-names(ndata[as.numeric(ndata)<4])
@@ -81,23 +73,18 @@ ddfilter.speed<-function (sdata, vmax=8.9, method=2){
       
       
     
-      # Speed from a previous and to a subsequent location in km/h
-      sdata$pSpeed<-sdata$pDist/sdata$pTime
-      sdata$sSpeed<-sdata$sDist/sdata$sTime
-    
-    
-      # Select locations at which the speed from a previous and to a subsequent location exceeds maximum linear traveling speed (Vmax)
+      # Select locations at which the distance from a previous and to a subsequent location exceeds maximum linear distance
       ## Function to identify location to remove: (0 = remove, 1 = keep)
       if(method==2){
         overMax<-function(i)
-        if(sdata$pSpeed[i]>vmax && sdata$sSpeed[i]>vmax && (!is.na(sdata$pSpeed[i])) && (!is.na(sdata$sSpeed[i]))){
+        if(sdata$pDist[i]>distance1 && sdata$sDist[i]>distance2 && (!is.na(sdata$pDist[i])) && (!is.na(sdata$sDist[i]))){
           0
         } else {
           1
         }
       } else if (method==1) {
         overMax<-function(i)
-        if((sdata$pSpeed[i]>vmax | sdata$sSpeed[i]>vmax) && (!is.na(sdata$pSpeed[i])) && (!is.na(sdata$sSpeed[i]))){
+        if((sdata$pDist[i]>distance1 | sdata$sDist[i]>distance2) && (!is.na(sdata$pDist[i])) && (!is.na(sdata$sDist[i]))){
           0
         } else {
           1
@@ -119,7 +106,7 @@ ddfilter.speed<-function (sdata, vmax=8.9, method=2){
       
       #### Bring back excluded data
       if(nrow(excluded.data)>0){
-        excluded.data[,c("pTime", "sTime", "pDist", "sDist", "pSpeed", "sSpeed", "overMax")]<-NA
+        excluded.data[,c("pTime", "sTime", "pDist", "sDist", "overMax")]<-NA
         sdata<-rbind(sdata, excluded.data)
       } else {
         sdata<-sdata
@@ -129,12 +116,12 @@ ddfilter.speed<-function (sdata, vmax=8.9, method=2){
 
   
   # Repeat the above function until no locations can be removed by this filter.
-  sdata2<-max.speed(sdata, vmax, method)
-  sdata3<-max.speed(sdata2, vmax, method)
+  sdata2<-max.distance(sdata, distance1, distance2, method)
+  sdata3<-max.distance(sdata2, distance1, distance2, method)
   while(!(nrow(sdata2) %in% nrow(sdata3)))
   {
-    sdata3<-max.speed(sdata2, vmax, method)
-    sdata2<-max.speed(sdata3, vmax, method)
+    sdata3<-max.distance(sdata2, distance1, distance2, method)
+    sdata2<-max.distance(sdata3, distance1, distance2, method)
   }
   
   
@@ -149,19 +136,18 @@ ddfilter.speed<-function (sdata, vmax=8.9, method=2){
   
   ## Print report
   if(length(id.exclude)>0){
-      cat("ddfilter.speed removed", RemovedSamplesN, "of", OriginalSS, "locations")
+      cat("distance_filter removed", RemovedSamplesN, "of", OriginalSS, "locations")
       cat("\n")
-      cat("  Warning: insufficient data to run ddfilter.speed for", id.exclude)
+      cat("  Warning: insufficient data to run distance_filter for", id.exclude)
       cat("\n")
   } else {
-      cat("ddfilter.speed removed", RemovedSamplesN, "of", OriginalSS, "locations")
+      cat("distance_filter removed", RemovedSamplesN, "of", OriginalSS, "locations")
       cat("\n")
   }
 
   
   # Delete working columns and return the output
-  sdata3$overMax<-NULL
-  # drops<-c("overMax", "sTime", "sDist", "pSpeed", "sSpeed")
-  # sdata3<-sdata3[,!(names(sdata3) %in% drops)] 
+  drops<-c("overMax", "sTime", "sDist")
+  sdata3<-sdata3[,!(names(sdata3) %in% drops)] 
   return(sdata3)
 }